Add a Contiki-NG base mote type
Put the common code for MspMoteType/ContikiMoteType
in a base class.

The code for compiling motes will also move into
this class, but that requires more than just
moving code so start with the simple parts.
